requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.5
daphne==3.0.1
distlib==0.3.1
Django==3.1.6
django-filter==2.4.0
djangochannelsrestframework==0.2.1
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
graphviz==0.16
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
importlib-metadata==3.4.0
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.7.0
msgpack==1.0.2
numpy==1.20.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
lxml==4.6.2
aiohttp==3.7.3